Move functions to common
